delete DistributedAtomSpace class
