+++ hyperon_das/api.py
@@ -3,7 +3,6 @@ from typing import Any, Dict, List, Tuple, Union
 
 from hyperon_das_atomdb import WILDCARD
 
-from hyperon_das.das import DistributedAtomSpace, QueryOutputFormat
 from hyperon_das.exceptions import DatabaseTypeException, MethodNotAllowed
 from hyperon_das.factory import DatabaseFactory, DatabaseType, database_factory
 from hyperon_das.logger import logger
@@ -11,9 +10,10 @@ from hyperon_das.pattern_matcher import (
     LogicalExpression,
     PatternMatchingAnswer,
 )
+from hyperon_das.utils import QueryOutputFormat
 
 
-class DistributedAtomSpaceAPI(DistributedAtomSpace):
+class DistributedAtomSpaceAPI:
     def __init__(self, database: DatabaseType) -> None:
         self._db_type = database
         try:
@@ -28,6 +28,49 @@ class DistributedAtomSpaceAPI(DistributedAtomSpace):
         logger().info(
             f"New Distributed Atom Space. Database name: {self.db.database_name}"
         )
+
+    def _to_handle_list(
+        self, atom_list: Union[List[str], List[Dict]]
+    ) -> List[str]:
+        if not atom_list:
+            return []
+        if isinstance(atom_list[0], str):
+            return atom_list
+        else:
+            return [handle for handle, _ in atom_list]
+
+    def _to_link_dict_list(
+        self, db_answer: Union[List[str], List[Dict]]
+    ) -> List[Dict]:
+        if not db_answer:
+            return []
+        flat_handle = isinstance(db_answer[0], str)
+        answer = []
+        for atom in db_answer:
+            if flat_handle:
+                handle = atom
+                arity = -1
+            else:
+                handle, targets = atom
+                arity = len(targets)
+            answer.append(self.db.get_atom_as_dict(handle, arity))
+        return answer
+
+    def _to_json(self, db_answer: Union[List[str], List[Dict]]) -> List[Dict]:
+        answer = []
+        if db_answer:
+            flat_handle = isinstance(db_answer[0], str)
+            for atom in db_answer:
+                if flat_handle:
+                    handle = atom
+                    arity = -1
+                else:
+                    handle, targets = atom
+                    arity = len(targets)
+                answer.append(
+                    self.db.get_atom_as_deep_representation(handle, arity)
+                )
+        return json.dumps(answer, sort_keys=False, indent=4)
 
     def clear_database(self) -> None:
         """Clear all data"""
