--- hyperon_das/das.py
@@ -1,54 +1,0 @@
-import json
-from enum import Enum, auto
-from typing import Dict, List, Union
-
-
-class QueryOutputFormat(int, Enum):
-    HANDLE = auto()
-    ATOM_INFO = auto()
-    JSON = auto()
-
-
-class DistributedAtomSpace:
-    def _to_handle_list(
-        self, atom_list: Union[List[str], List[Dict]]
-    ) -> List[str]:
-        if not atom_list:
-            return []
-        if isinstance(atom_list[0], str):
-            return atom_list
-        else:
-            return [handle for handle, _ in atom_list]
-
-    def _to_link_dict_list(
-        self, db_answer: Union[List[str], List[Dict]]
-    ) -> List[Dict]:
-        if not db_answer:
-            return []
-        flat_handle = isinstance(db_answer[0], str)
-        answer = []
-        for atom in db_answer:
-            if flat_handle:
-                handle = atom
-                arity = -1
-            else:
-                handle, targets = atom
-                arity = len(targets)
-            answer.append(self.db.get_atom_as_dict(handle, arity))
-        return answer
-
-    def _to_json(self, db_answer: Union[List[str], List[Dict]]) -> List[Dict]:
-        answer = []
-        if db_answer:
-            flat_handle = isinstance(db_answer[0], str)
-            for atom in db_answer:
-                if flat_handle:
-                    handle = atom
-                    arity = -1
-                else:
-                    handle, targets = atom
-                    arity = len(targets)
-                answer.append(
-                    self.db.get_atom_as_deep_representation(handle, arity)
-                )
-        return json.dumps(answer, sort_keys=False, indent=4)
