+++ hyperon_das/utils.py
@@ -1,0 +1,8 @@
+import json
+from enum import Enum, auto
+
+
+class QueryOutputFormat(int, Enum):
+    HANDLE = auto()
+    ATOM_INFO = auto()
+    JSON = auto()
